Fix exceptions caused by tasks that fall "between" the start of the sprint and the first day of the sprint.
An exception was introduced by the last commit. This is an example:

Start date is Sep 10 at 6:00PM
A task was created, and closed sometime between then and midnight.

The events did not happen prior to the start_date, so they wouldn't get
classified as "before start of sprint", but when you convert that date to
a string for storing in the array, it gets stored as "Thu Sep 10" (for
example) which does not exist in the array, causing an exception.
Timezones further complicated things.

In retrospect, the sprint start and end probably should have been dates,
not date times. Although then I'm not sure how it would behave in
different time zones.

In reality, this data structure should be flushed out better, and more
thought be given to timezones. Should the first "day" of a sprint be the
first 24 hours after the start datetime, resulting in a possible partial
day at the beggining of the sprint. Or could there be partial days at the
beginning and end of the sprint? Should the start and end dates just be
dates and the sprints start and end at midnight based on your time zone?